requirements.txt
# V3 Capstone Dependencies

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# ML/Embeddings
scikit-learn>=1.0.0
sentence-transformers>=2.2.0

# LangChain for Traditional RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
networkx>=3.0

# Evaluation (official RAGAS)
ragas>=0.1.0
langchain-openai>=0.1.0

# Optional: for faster vector operations
# faiss-cpu>=1.7.0
